+++ src/Fakes/v1/YousignFaker.php
@@ -5,6 +5,7 @@ namespace Coverzen\Components\YousignClient\Fakes\v1;
 use Closure;
 use Coverzen\Components\YousignClient\Libs\Soa\v1\Soa;
 use Coverzen\Components\YousignClient\Libs\Soa\v1\Yousign;
+use Coverzen\Components\YousignClient\Structs\Soa\v1\AddConsentResponse;
 use Coverzen\Components\YousignClient\Structs\Soa\v1\InitiateSignatureResponse;
 use Coverzen\Components\YousignClient\Structs\Soa\v1\UploadDocumentResponse;
 use Coverzen\Components\YousignClient\YousignClientServiceProvider;
@@ -85,6 +86,12 @@ class YousignFaker
                     UploadDocumentResponse::factory()
                                           ->make()
                                           ->toArray(),
+                    Response::HTTP_CREATED
+                ),
+                $url . Yousign::INITIATE_SIGNATURE_URL . '/*/' . Yousign::ADD_CONSENT_URL => Http::response(
+                    AddConsentResponse::factory()
+                                      ->make()
+                                      ->toArray(),
                     Response::HTTP_CREATED
                 ),
             ]