+++ src/Libs/Soa/v1/Yousign.php
@@ -4,6 +4,8 @@ namespace Coverzen\Components\YousignClient\Libs\Soa\v1;
 
 use Coverzen\Components\YousignClient\Structs\Soa\v1\AddSignerRequest;
 use Coverzen\Components\YousignClient\Structs\Soa\v1\AddSignerResponse;
+use Coverzen\Components\YousignClient\Structs\Soa\v1\AddConsentRequest;
+use Coverzen\Components\YousignClient\Structs\Soa\v1\AddConsentResponse;
 use Coverzen\Components\YousignClient\Structs\Soa\v1\InitiateSignatureRequest;
 use Coverzen\Components\YousignClient\Structs\Soa\v1\InitiateSignatureResponse;
 use Coverzen\Components\YousignClient\Structs\Soa\v1\UploadDocumentRequest;
@@ -16,6 +18,7 @@ use Illuminate\Support\Arr;
 use Illuminate\Support\Facades\Http;
 use Illuminate\Support\Str;
 use RuntimeException;
+use function implode;
 use function is_array;
 
 /**
@@ -28,6 +31,9 @@ class Yousign extends Soa
 
     /** @var string */
     public const UPLOAD_DOCUMENT_URL = 'documents';
+
+    /** @var string */
+    public const ADD_CONSENT_URL = 'consent_requests';
 
     /** @var string */
     public const ADD_SIGNER_URL = 'signers';
@@ -96,9 +102,6 @@ class Yousign extends Soa
             throw new RuntimeException('File content is required.');
         }
 
-        /** @var string $url */
-        $url = self::INITIATE_SIGNATURE_URL . self::URL_SEPARATOR . $signatureRequestId . self::URL_SEPARATOR . self::UPLOAD_DOCUMENT_URL;
-
         /** @var Response $response */
         $response = $this->apiClient->attach(
             self::FILE_PARAM,
@@ -106,7 +109,14 @@ class Yousign extends Soa
             $uploadDocumentRequest->file_name,
         )
                                     ->post(
-                                        $url,
+                                        implode(
+                                            self::URL_SEPARATOR,
+                                            [
+                                                self::INITIATE_SIGNATURE_URL,
+                                                $signatureRequestId,
+                                                self::UPLOAD_DOCUMENT_URL,
+                                            ]
+                                        ),
                                         [
                                             self::NATURE_PARAM => $uploadDocumentRequest->nature->value,
                                         ]
@@ -138,5 +148,33 @@ class Yousign extends Soa
         }
 
         return new AddSignerResponse($response->json());
+    }
+
+    /**
+     * @param string $signatureRequestId
+     * @param AddConsentRequest $addConsentRequest
+     *
+     * @return AddConsentResponse
+     */
+    public function addConsent(string $signatureRequestId, AddConsentRequest $addConsentRequest): AddConsentResponse
+    {
+        /** @var Response $response */
+        $response = $this->apiClient->post(
+            implode(
+                self::URL_SEPARATOR,
+                [
+                    self::INITIATE_SIGNATURE_URL,
+                    $signatureRequestId,
+                    self::ADD_CONSENT_URL,
+                ]
+            ),
+            $addConsentRequest->toArray()
+        );
+
+        if (!is_array($response->json())) {
+            throw new RuntimeException('Yousign response is not an array.');
+        }
+
+        return new AddConsentResponse($response->json());
     }
 }